third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-tables/tentative/baseline-td-expected.txt

This is a testharness.js-based test.
[FAIL] table, .display-table 1
  assert_equals: \n<table style="height:100px" class="ahem vgrad">\n  <tbody><tr><td style="font: 75px/1 Ahem;vertical-align:baseline;"><div class="m" data-offset-y="0">Xp</div></td>\n  <td style="" title="default"><div class="m" data-offset-y="25">Xp</div></td>\n  <td style="vertical-align: top"><div class="m" data-offset-y="0">Xp</div></td>\n  <td style="vertical-align: middle"><div class="m" data-offset-y="25">Xp</div></td>\n  <td style="vertical-align: bottom"><div class="m" data-offset-y="50">Xp</div></td>\n  <td style="vertical-align: baseline"><div class="m" data-offset-y="28">Xp</div></td>\n  <td style="vertical-align: sub"><div class="m" data-offset-y="28">Xp</div></td>\n  <td style="vertical-align: super"><div class="m" data-offset-y="28">Xp</div></td>\n  <td style="vertical-align: text-top"><div class="m" data-offset-y="28">Xp</div></td>\n  <td style="vertical-align: text-bottom"><div class="m" data-offset-y="28">Xp</div></td>\n  <td style="vertical-align: 100px"><div class="m" data-offset-y="28">Xp</div></td>\n  <td style="vertical-align: 100%"><div class="m" data-offset-y="28">Xp</div></td>\n</tr></tbody></table>\noffsetTop expected 28 but got 30
[FAIL] table, .display-table 2
  assert_equals: \n<div style="height:100px;position:relative" class="ahem vgrad display-table">\n  <div class="display-cell" style="font: 75px/1 Ahem;vertical-align:baseline;"><div class="m" data-offset-y="0">Xp</div></div>\n  <div class="display-cell" style="" title="default"><div class="m" data-offset-y="28">Xp</div></div>\n  <div class="display-cell" style="vertical-align: top"><div class="m" data-offset-y="0">Xp</div></div>\n  <div class="display-cell" style="vertical-align: middle"><div class="m" data-offset-y="25">Xp</div></div>\n  <div class="display-cell" style="vertical-align: bottom"><div class="m" data-offset-y="50">Xp</div></div>\n  <div class="display-cell" style="vertical-align: baseline"><div class="m" data-offset-y="28">Xp</div></div>\n  <div class="display-cell" style="vertical-align: sub"><div class="m" data-offset-y="28">Xp</div></div>\n  <div class="display-cell" style="vertical-align: super"><div class="m" data-offset-y="28">Xp</div></div>\n  <div class="display-cell" style="vertical-align: text-top"><div class="m" data-offset-y="28">Xp</div></div>\n  <div class="display-cell" style="vertical-align: text-bottom"><div class="m" data-offset-y="28">Xp</div></div>\n  <div class="display-cell" style="vertical-align: 100px"><div class="m" data-offset-y="28">Xp</div></div>\n  <div class="display-cell" style="vertical-align: 100%"><div class="m" data-offset-y="28">Xp</div></div>\n</div>\noffsetTop expected 28 but got 30
Harness: the test ran to completion.
